+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The public API of this library consists of the public functions declared in
 file [H3Core.java](./src/main/java/com/uber/h3core/H3Core.java).
 
 ## [Unreleased]
+### Added
+- Added script for cross compiling using dockcross.
+- Changed the native library loader to detect more operating systems.
 
 ## [3.0.0] - 2018-03-27
 ### Added